requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0
pytest>=8.0.0